requirements.txt
jsons==1.6.3
requests==2.32.5
telebot==0.0.5
python-dotenv~=1.2.1
typing~=3.7.4.3